requirements.txt
# Core dependencies
dolphin-memory-engine==1.3.0
colorama==0.4.6
flask==2.3.3
python-dotenv==1.0.0
pyautogui==0.9.54
pillow>=10.0.0
pygetwindow==0.0.9
opencv-python==4.8.0.76
mss==9.0.1
pywin32==310
keyboard>=0.13.5
redis==5.0.1
flask-socketio==5.3.6
python-socketio==5.11.0
openai==1.5.0

# OmniParser dependencies
numpy==1.26.4  # Version fixe pour éviter NumPy 2.x
torch>=2.3.0
torchvision>=0.18.0
transformers==4.49.0  # Version spécifique pour Florence-2
ultralytics>=8.0.0  # YOLO
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
huggingface-hub>=0.16.0
onnxruntime==1.15.1  # Version compatible avec NumPy 1.x
